fixed double to a long long, opps

+++ 0x00-hello_world/6-size.c
@@ -12,7 +12,7 @@ int main(void)
 	char a;
 	int b;
 	long c;
-	double d;
+	long long d;
 	float e;
 
 	printf("Size of a char: %d bytes(s)\n", sizeof(a));
